Sites list: Display only icons as site thumbnails
